refactor: Phase 2 — extract shared utilities
- src/utils/formatting.ts: consolidate formatCurrency, formatCurrencyShort,
  formatDate, toSafeNumber (previously duplicated across Dashboard, Transactions,
  Charts with slight variations)
- src/utils/memo.ts: extract memoizeByReference (was duplicated in aggregation.ts
  and filters.ts)
- date-utils.ts: export toMonthKey and toDateKey (were private duplicates in
  aggregation.ts and chart-data.ts)

All three components and both services now import from the shared utils.
635 tests passing, lint clean.
