+++ src/utils/date-utils.ts
@@ -209,3 +209,16 @@ export function generatePeriodOptions(
 
   return options;
 }
+
+export function toDateKey(date: Date): string {
+  const year = date.getUTCFullYear()
+  const month = `${date.getUTCMonth() + 1}`.padStart(2, '0')
+  const day = `${date.getUTCDate()}`.padStart(2, '0')
+  return `${year}-${month}-${day}`
+}
+
+export function toMonthKey(date: Date): string {
+  const year = date.getUTCFullYear()
+  const month = `${date.getUTCMonth() + 1}`.padStart(2, '0')
+  return `${year}-${month}`
+}
